dom/base/test/test_bug1126851.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1126851
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1126851</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1126851 **/

  SimpleTest.waitForExplicitFinish();

  function runTest() {
    win = window.open("about:blank", "");
    win.onload = function() {
      win.onunload = function() {
        ok(true, "got unload");
        win.close();
        SimpleTest.finish();
      }
      win.onresize = function() {
        win.location.reload();
      }
      win.document.dispatchEvent(new win.Event("resize", { bubbles: true }));
    }
  }


  </script>
</head>
<body onload="runTest()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1126851">Mozilla Bug 1126851</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
